requirements.txt
Flask==2.0.1
python-dotenv==1.0.1
supabase==2.3.4
python-jose==3.3.0
gunicorn==20.1.0 